+++ Input/Input.js
@@ -1,4 +1,5 @@
 import {handle, forKey, forward, forwardCustom} from '@enact/core/handle';
+import deprecate from '@enact/core/internal/deprecate';
 import kind from '@enact/core/kind';
 import {extractAriaProps} from '@enact/core/util';
 import Spotlight from '@enact/spotlight';
@@ -66,6 +67,16 @@ const InputPopupBase = kind({
 		css: PropTypes.object,
 
 		/**
+		 * Initial value of the input.
+		 *
+		 * This value is used for setting the `defaultValue` of the `InputField`.
+		 * @see {@link sandstone/Input.InputField}
+		 * @type {String|Number}
+		 * @public
+		 */
+		defaultValue: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+
+		/**
 		 * Disables the input popup.
 		 *
 		 * @type {Boolean}
@@ -82,7 +93,7 @@ const InputPopupBase = kind({
 		inputFieldSpotlightId: PropTypes.string,
 
 		/**
-		 * Indicates {@link sandstone/Input.InputPopupBase.value|value} is invalid and shows
+		 * Indicates the value is invalid and shows
 		 * {@link sandstone/Input.InputPopupBase.invalidMessage|invalidMessage}, if set.
 		 *
 		 * @type {Boolean}
@@ -289,11 +300,13 @@ const InputPopupBase = kind({
 		 * @see {@link sandstone/Input.InputField}
 		 * @type {String|Number}
 		 * @public
+		 * @deprecated Will be removed in 3.0.0. Use {@link sandstone/Input.InputPopupBase.defaultValue|defaultValue} instead.
 		 */
 		value: PropTypes.oneOfType([PropTypes.string, PropTypes.number])
 	},
 
 	defaultProps: {
+		defaultValue: '',
 		popupType: 'fullscreen',
 		numberInputField: 'auto',
 		size: 'small',
@@ -349,6 +362,7 @@ const InputPopupBase = kind({
 		backButtonAriaLabel,
 		children,
 		css,
+		defaultValue,
 		inputFieldSpotlightId,
 		noBackButton,
 		noSubmitButton,
@@ -372,6 +386,13 @@ const InputPopupBase = kind({
 		minLength,
 		...rest
 	}) => {
+		if (value) {
+			deprecate({
+				name: 'sandstone/Input.InputPopupBase.value',
+				replacedBy: 'sandstone/Input.InputPopupBase.defaultValue',
+				until: '3.0.0'
+			});
+		}
 		const id = `inputPopup`;
 		const ariaLabelledBy = popupAriaLabel ? null : `${id}_title ${id}_subtitle`;
 		const inputProps = extractInputFieldProps(rest);
@@ -427,7 +448,7 @@ const InputPopupBase = kind({
 									announce={announce}
 									maxLength={limitNumberLength(popupType, maxLength)}
 									minLength={limitNumberLength(popupType, minLength)}
-									defaultValue={value}
+									defaultValue={defaultValue || value}
 									onBeforeChange={onBeforeChange}
 									onComplete={onNumberComplete}
 									showKeypad
@@ -444,7 +465,7 @@ const InputPopupBase = kind({
 									size={size}
 									autoFocus
 									type={type}
-									defaultValue={value}
+									defaultValue={defaultValue || value}
 									placeholder={placeholder}
 									onBeforeChange={onBeforeChange}
 									onKeyDown={onInputKeyDown}
@@ -479,6 +500,15 @@ const InputBase = kind({
 		 * @private
 		 */
 		announce: PropTypes.func,
+
+		/**
+		 * Initial value of the input.
+		 *
+		 * @see {@link sandstone/Input.InputField}
+		 * @type {String|Number}
+		 * @public
+		 */
+		defaultValue: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 
 		/**
 		 * Disables the button that activates the input popup.
@@ -516,9 +546,8 @@ const InputBase = kind({
 		type: PropTypes.oneOf(['text', 'password', 'number', 'passwordnumber', 'url', 'tel', 'passwordtel']),
 
 		/**
-		 * Initial value of the input.
+		 * The value of the input.
 		 *
-		 * This value is used for setting the `defaultValue` of the `InputField`.
 		 * @see {@link sandstone/Input.InputField}
 		 * @type {String|Number}
 		 * @public
@@ -566,7 +595,7 @@ const InputBase = kind({
 					type={type}
 					size={size}
 					disabled={disabled}
-					value={value}
+					defaultValue={value}
 					placeholder={placeholder}
 					{...rest}
 				/>